refactor: split ml-telemetry.ts into types, validators, and aggregators
